# lib/api-zod/src/generated/types/types/contactMethodInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMethodInput {
  /** @minLength 1 */
  channelType: string;
  /** @minLength 1 */
  channelValue: string;
  isPreferred?: boolean;
}