# lib/api-zod/src/generated/types/types/contactWithMethods.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from './contact';
import type { ContactMethod } from './contactMethod';

export type ContactWithMethods = Contact & {
  methods: ContactMethod[];
};