# lib/api-zod/src/generated/types/types/notificationLog.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NotificationLog {
  id: number;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  contactId?: number | null;
  channelType: string;
  channelValue: string;
  templateId: string;
  subject?: string;
  body?: string;
  status: string;
  /** @nullable */
  errorMessage?: string | null;
  sentAt: Date;
  /**
     * Baileys message key ID used to match delivery/read receipts
     * @nullable
     */
  whatsappMessageId?: string | null;
  /**
     * WhatsApp receipt status: delivered | read (null = not yet acknowledged)
     * @nullable
     */
  deliveryStatus?: string | null;
}