# lib/api-zod/src/generated/types/types/systemStatusInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SystemStatusInput {
  /** @minLength 1 */
  platform: string;
  status?: string;
  notes?: string;
}