# lib/api-zod/src/generated/types/types/systemStatusRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SystemStatusRecord {
  id: number;
  platform: string;
  status: string;
  lastChecked: Date;
  /** @nullable */
  lastSync?: Date | null;
  notes?: string;
  createdAt: Date;
}